refactor: move `DefaultReference` to `app/models`
