+++ app/controllers/my/default_reference_controller.rb
@@ -4,7 +4,7 @@ module My
     before_action :set_reference, only: :update
 
     def update
-      References::DefaultReference.set session, @reference
+      DefaultReference.set session, @reference
 
       redirect_back fallback_location: references_path,
         notice: "#{@reference.keey} was successfully set as the default reference."
